# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
#host: localhost
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /umf:
    x-swagger-router-controller: epivector.umf.controller
    
    get:
      operationId: getUmf
      description: Retorna la lista de entidades
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UmfList"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"          
                      
    post:
      description: Agrega una nueva entidad
      parameters:
        - in: body
          name: entidad
          schema:
            $ref: "#/definitions/Umf"
      operationId: createUmf
      responses: 
        "201":
          description: success
          schema:
            $ref: "#/definitions/UmfResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
  /umfs/{id}:
    x-swagger-router-controller: epivector.umf.controller
    
    get: 
      description: Retorna una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path
          required: true
      operationId: getUmfId
      responses:
        "200":
          description: success
          schema:
           $ref: "#/definitions/UmfResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
            
    put:
      description: Actualiza una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path
          required: true
        - in: body
          name: entidad
          schema:
            $ref: "#/definitions/Umf"
      operationId: updateUmf
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UmfResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      description: Borra una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path 
          required: true
      operationId: deleteUmf
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      code:
        type: number
      message:
        type: string
      description:
        type: string
  #  GeneralResponse:
  #    type: object
  #    properties:
  #      success:
  #        type: number
  #        description: returns 1 if successful
  #      description:
  #        type: string
  #        description: a short comment
  #    required:
  #      - success
  #      - description
...
  DeleteResponse:
    type: object
    properties:
      success:
        type: number
        description: Retorna 1 en caso de exito
      description:
        type: string
        description: Comentario
    required:
      - success
      - description
...
  Umf:
    type: object
    properties:
      instalación:
        type: string
        description: el tipo de instancia
      tipo_de_instalación: 
        type: string
        description: tipo de instancia familiar
      tipo_de_unidad:
        type: string
        description: tipo de unidad medica
      dirección:
        type: string
        description: direccion de la unidad umf
      entre_calles:
        type: string
        description: direccion entre calles
      teléfono:
        type: number
        description: telefono de la instancia
      horario_de_atención:
        type: string
        description: horario de atencion
  UmfResponse:
    type: object
    properties:
      instalación:
        type: string
        description: el tipo de instancia
      tipo_de_instalación: 
        type: string
        description: tipo de instancia familiar
      tipo_de_unidad:
        type: string
        description: tipo de unidad medica
      dirección:
        type: string
        description: direccion de la unidad umf
      entre_calles:
        type: string
        description: direccion entre calles
      teléfono:
        type: number
        description: telefono de la instancia
      horario_de_atención:
        type: string
        description: horario de atencion
  UmfList:
    properties:
      umf:
        type: array
        items:
          $ref: "#/definitions/UmfResponse"
    required:
      - umf